pyproject.toml
[tool.poetry]
name = "finsee"
version = "0.3.0"
description = "Financial Notes both income and outcome"
authors = ["ahmadsyafrudin <sfrmd4@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1"
honcho = "^1.0.1"
psycopg2-binary = "^2.8.5"
djangorestframework = "^3.11.1"
bump2version = "^1.0.0"

[tool.poetry.dev-dependencies]
coverage = "^5.2.1"
pytest = "^6.0.1"
black = "^19.10b0"
pre-commit = "^2.6.0"
honcho = "^1.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
